Remove unnecessary function ClientImpl::getLine().
The function ClientImpl::getLine() is only used at one place in the
whole library. It therefore doesn't seem to be really useful to keep it.

Even though a possible argument for keeping it may be that it abstracts
away getting a single line, I don't think that is a really strong
argument.

Because this function is only used once, and in a function that directly
uses socket functions itself already, this makes the code more complex
than needed.